Update telemetry entry sh to support k8s start and keep compatible with systemd start flow
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
Update telemetry entry sh to support k8s start and keep compatible with systemd start flow

#### How I did it
With k8s start, it will contains below env

sudo docker run -d --rm -it --privileged \
  --net=host --pid=host --ipc=host \
  --tmpfs /usr/share/sonic:rw \
  -v /usr/share/sonic/device:/usr/share/sonic/device:ro \
  -v /etc/sonic:/etc/sonic:ro \
  -v /etc/localtime:/etc/localtime:ro \
  -v /host/reboot-cause:/host/reboot-cause \
  -v /var/run/redis:/var/run/redis \
  -v /var/run/redis-chassis:/var/run/redis-chassis:ro \
  -v /usr/share/sonic/templates/rsyslog-container.conf.j2:/usr/share/sonic/templates/rsyslog-container.conf.j2:ro \
  -e RUNTIME_OWNER=kube \
  -e SYSLOG_TARGET_IP=127.0.0.1 \
  -e NAMESPACE_PREFIX=asic \
  -e launch_by=k8s \
  docker-sonic-telemetry:latest

by default it will not be used so that existing rollout mgmt test will not be break.

#### How to verify it
<img width="1293" height="336" alt="image" src="https://github.com/user-attachments/assets/474b69e7-13ad-47c1-b1ff-52653a203b75" />

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)