[Smartswitch][Mellanox]Change sensor label for DPUs
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
Two of the DPUs had incorrect labels in some of the sensors in the sensors configuration which is consumed by sensord. This pr fixes the name for these specific DPUs

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
By changing the label in sensors.conf in sn4280 platform file
Change the label to indicate DPU4 in sensors.conf

#### How to verify it
Manually check the sensors output to confirm that DPU4 contains only DPU4 related sensors
Without fix
```
root@sonic:/home/admin# jq '.["mp2975-i2c-21-69"]["PMIC-12 12V DPU3 VDD VDDQ_DDR (in)"]' a.json
{
  "power1_input": 32,
  "power1_alarm": 0
}
```
With fix
```
root@sonic:/home/admin# jq '.["mp2975-i2c-21-69"]["PMIC-12 12V DPU4 VDD VDDQ_DDR (in)"]' a.json
{
  "power1_input": 31,
  "power1_alarm": 0
}
```

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)