[Nvidia-bluefield][DPU]Fixed allowing PF interface to update MAC address
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it

Previously the PF interface MAC address could not be set directly due to driver issues, now the need for a bridge network is not present since we can set the PF interface MAC address directly.
Therefore the additional service and bridge network is not required since the mac address can be set directly using ip link command

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
Removed `pf-manager.service` This service was responsible to configure the pf0 interface as part of the eth0-midplane bridge and set the mac for eth0-midplane interface
Changed configure_midplane_iface command to only set the ip address of the midplane interface
Changed the udev rule to call the configure_midplane_iface function directly instead of using a service

#### How to verify it
Manual check with ifconfig the confirm that the changes are in effect (ifconfig command to confirm the ip address is updated)
Run tests to confirm that there are no regressions

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)