[Arista][DNX] Workaround for sai.profile being empty when syncd starts (#1770)
We're seeing an issue where start.sh is run multiple times on syncd
startup and if the 'cp' operation in the 2nd start.sh occurs while the
Syncd process is reading the /etc/sai.d/sai.profile it will end up
reading an empty file.

Therefor if we skip the 'cp' operation when the file is unchanged the
2nd start.sh run should always be a no-op.

This is a workaround until we find the reason and solution to stop
start.sh from running multiple times.